cruft removal
